pyproject.toml
[tool.poetry]
name = "jsac2023"
version = "0.1.0"
description = ""
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cvss = ">=2.5,<3.0"
cyclonedx-bom = ">=3.7,<4.0"
httpx = ">=0.23,<1.0"
loguru = ">=0.6,<1.0"
pip-audit = ">=2.4,<3.0"
psutil = ">=5.9,<6.0"
pydantic = ">=1.10,<2.0"
pyhumps = ">=3.8,<4.0"
pyparsing = ">=3.0,<4.0"
sh = ">=1.14,<2.0"
typer = ">=0.7,<1.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
mypy = "^0.991"
pre-commit = "^2.20.0"
pycodestyle = "^2.9.1"
pytest = "^7.2.0"
vcrpy = "^4.2.1"
types-setuptools = "^65.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
jsac2023-cli = "jsac2023.cli:app"
